pyproject.toml
[project]
name = "microhubs-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]==0.128.0",
    "uvicorn[standard]==0.40.0",
    "geojson-pydantic==1.1.0",
    "psycopg2==2.9.10",
    "PyJWT==2.3.0",
    "redis==7.1.0",
    "simplekml==1.3.6",
    "shapely==2.1.2",
    "python-multipart==0.0.21",
    "fastkml==0.12",
    "setuptools==75.4.0",
    "pytz==2024.1",
    "fudgeo==0.8.0",
    "isodate==0.7.2",
    "pylint>=4.0.4",
]
